Refactor marker management: move MarkerContext and related components to a new 'marker' directory
