+++ electron/src/components/graph/marker/useMarkerManager.ts
